Ran SPLINT (secure programming LINT) and fixed warnings.

+++ trunk/bacnet-stack/src/bacdcode.c
@@ -1193,7 +1193,7 @@ int encode_context_unsigned(
         len = 4;
     }
 
-    len = encode_tag(&apdu[0], tag_number, true, len);
+    len = encode_tag(&apdu[0], tag_number, true, (uint32_t)len);
     len += encode_bacnet_unsigned(&apdu[len], value);
 
     return len;
@@ -1566,7 +1566,7 @@ int decode_bacnet_time_safe(
         btime->hundredths = 0;
         btime->min = 0;
         btime->sec = 0;
-        return len_value;
+        return (int)len_value;
     } else {
         return decode_bacnet_time(apdu, btime);
     }
@@ -1699,7 +1699,7 @@ int decode_date_safe(
         bdate->month = 0;
         bdate->wday = 0;
         bdate->year = 0;
-        return len_value;
+        return (int)len_value;
     } else {
         return decode_date(apdu, bdate);
     }
