+++ trunk/bacnet-stack/src/ptransfer.c
@@ -88,7 +88,7 @@ int ptransfer_encode_apdu(
         apdu[3] = SERVICE_CONFIRMED_PRIVATE_TRANSFER;
         apdu_len = 4;
         apdu_len =
-            pt_encode_apdu(&apdu[apdu_len], MAX_APDU - apdu_len, private_data);
+            pt_encode_apdu(&apdu[apdu_len], (uint16_t)(MAX_APDU - apdu_len), private_data);
     }
 
     return apdu_len;
@@ -105,7 +105,7 @@ int uptransfer_encode_apdu(
         apdu[1] = SERVICE_UNCONFIRMED_PRIVATE_TRANSFER;
         apdu_len = 2;
         apdu_len =
-            pt_encode_apdu(&apdu[apdu_len], MAX_APDU - apdu_len, private_data);
+            pt_encode_apdu(&apdu[apdu_len], (uint16_t)(MAX_APDU - apdu_len), private_data);
     }
 
     return apdu_len;
